+++ plume/src/main/kotlin/io/github/plume/oss/domain/model/DeltaGraph.kt
@@ -2,6 +2,11 @@ package io.github.plume.oss.domain.model
 
 import io.github.plume.oss.drivers.IDriver
 import io.shiftleft.codepropertygraph.generated.nodes.NewNodeBuilder
+import overflowdb.Config
+import overflowdb.Graph
+import overflowdb.Node
+import io.shiftleft.codepropertygraph.generated.edges.Factories as EdgeFactories
+import io.shiftleft.codepropertygraph.generated.nodes.Factories as NodeFactories
 
 /**
  * This is based off of [io.shiftleft.passes.DiffGraph]. Where DiffGraphs do not propagate ID changes, this one does by
@@ -27,6 +32,43 @@ class DeltaGraph private constructor(private val changes: List<Delta>) {
             }
         }
     }
+
+    /**
+     * Applies the delta graph to an OverflowDB instance. To have valid IDs this must be passed into the driver with
+     * the [DeltaGraph.apply] method first.
+     */
+    fun toOverflowDb(): Graph =
+        Graph.open(
+            Config.withDefaults(),
+            NodeFactories.allAsJava(),
+            EdgeFactories.allAsJava()
+        ).let { g ->
+            fun addNode(n: NewNodeBuilder): Node {
+                val b = n.build()
+                val v = if (n.id() > 0) g.addNode(n.id(), b.label()) else g.addNode(b.label())
+                b.properties().foreachEntry { key, value -> v.setProperty(key, value) }
+                n.id(v.id())
+                return v
+            }
+            changes.forEach { d ->
+                when (d) {
+                    is VertexAdd -> addNode(d.n)
+                    is VertexDelete -> g.node(d.id)?.remove()
+                    is EdgeAdd -> {
+                        val src = if (g.node(d.src.id()) != null) g.node(d.src.id()) else addNode(d.src)
+                        val dst = if (g.node(d.dst.id()) != null) g.node(d.dst.id()) else addNode(d.dst)
+                        src.addEdge(d.e, dst)
+                    }
+                    is EdgeDelete -> {
+                        val src = g.node(d.src.id())
+                        g.node(d.dst.id())?.let { dst ->
+                            src.outE(d.e).asSequence().firstOrNull { it.inNode() == dst }?.remove()
+                        }
+                    }
+                }
+            }
+            return g
+        }
 
     /**
      * Builds an [DeltaGraph] instance by accumulating changes.